requirements.txt
flask>=2.3.0
biopython>=1.81
matplotlib>=3.7.0
moviepy>=1.0.3
numpy>=1.24.0
werkzeug>=2.3.0
gunicorn>=21.0.0 